fix: exact band heights — no hairline of header above the pinned tab bar
The shell rounded the measured header/tab-bar heights to whole dp before
handing them to native (Int32 props). With the bands stacked inside one
scroll view since 0.8.0, a 223.33 pt header collapsed by 223 left 0.33 pt
of itself showing above the tab bar. headerHeight/tabBarHeight are Float
now, pages pad by the exact sum, and Android rounds dp→px the way Yoga
snaps layout to the pixel grid.

+++ android/src/main/java/com/collapsibletabs/bridge/CollapsibleTabsViewManager.kt
@@ -102,12 +102,12 @@ class CollapsibleTabsViewManager(
     // MARK: - Props
 
     @ReactProp(name = "headerHeight")
-    override fun setHeaderHeight(view: CollapsibleTabsHostView, value: Int) {
+    override fun setHeaderHeight(view: CollapsibleTabsHostView, value: Float) {
         view.setHeaderHeightDp(value)
     }
 
     @ReactProp(name = "tabBarHeight")
-    override fun setTabBarHeight(view: CollapsibleTabsHostView, value: Int) {
+    override fun setTabBarHeight(view: CollapsibleTabsHostView, value: Float) {
         view.setTabBarHeightDp(value)
     }
